+++ test/OpenTelemetry.Shims.OpenTracing.Tests/ListenAndSampleAllActivitySourcesFixture.cs
@@ -1,0 +1,30 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using System.Diagnostics;
+
+namespace OpenTelemetry.Shims.OpenTracing.Tests;
+
+public sealed class ListenAndSampleAllActivitySourcesFixture : IDisposable
+{
+    private readonly ActivityListener listener;
+
+    public ListenAndSampleAllActivitySourcesFixture()
+    {
+        Activity.DefaultIdFormat = ActivityIdFormat.W3C;
+        Activity.ForceDefaultIdFormat = true;
+
+        this.listener = new ActivityListener
+        {
+            ShouldListenTo = _ => true,
+            Sample = (ref ActivityCreationOptions<ActivityContext> options) => ActivitySamplingResult.AllDataAndRecorded,
+        };
+
+        ActivitySource.AddActivityListener(this.listener);
+    }
+
+    public void Dispose()
+    {
+        this.listener.Dispose();
+    }
+}
